Fixed "warning C4311: 'type cast': pointer truncation from 'HINSTANCE' to 'int'" in sys_posix.cpp on Windows (#9797)

+++ engine/dlib/src/dlib/sys_posix.cpp
@@ -274,7 +274,7 @@ namespace dmSys
 
     Result OpenURL(const char* url, const char* target)
     {
-        int ret = (int) ShellExecuteA(NULL, "open", url, NULL, NULL, SW_SHOWNORMAL);
+        intptr_t ret = (intptr_t) ShellExecuteA(NULL, "open", url, NULL, NULL, SW_SHOWNORMAL);
         if (ret == 32)
         {
             return RESULT_OK;
